extract out common sig public signal code
