+++ frontend/pages/api/submit.ts
@@ -3,7 +3,10 @@ import { ethers } from "ethers";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { PointPreComputes } from "../../types/zk";
 import { SECP256K1_N } from "../../utils/config";
-import { getPointPreComputes } from "../../utils/wasmPrecompute";
+import {
+  getSigPublicSignals,
+  getPointPreComputes,
+} from "../../utils/wasmPrecompute";
 
 import { prisma } from "../../utils/prisma";
 
@@ -26,39 +29,16 @@ export async function verifyProof(
   return proofVerified;
 }
 
-// NOTE: this code is duplicated with that in generateProof.tsx and should be extracted out
 async function verifySignatureArtifacts(
   publicSignals: string[],
   r: string,
-  isRYOdd: boolean,
+  isRYOdd: bigint,
   msg: string
 ): Promise<boolean> {
-  const rPoint = ec.keyFromPublic(
-    ec.curve.pointFromX(new BN(r.substring(2), 16), isRYOdd).encode("hex"),
-    "hex"
-  );
-
-  // Get the group element: -(m * r^−1 * G)
-  const rInv = new BN(r.substring(2), 16).invm(SECP256K1_N);
-
-  // w = -(r^-1 * msg)
-  const w = rInv
-    .mul(new BN(ethers.utils.hashMessage(msg).substring(2), 16))
-    .neg()
-    .umod(SECP256K1_N);
-
-  // U = -(w * G) = -(r^-1 * msg * G)
-  const U = ec.curve.g.mul(w);
-
-  // T = r^-1 * R
-  const T = rPoint.getPublic().mul(rInv);
-
-  const TPreComputes = await getPointPreComputes(T.encode("hex"));
+  const { TPreComputes, U } = await getSigPublicSignals(r, isRYOdd, msg);
 
   // TODO: compare epxected with TPreComputes in publicSignals (need to deserialize from strings)
-
   // TODO: compare expectedU with U in publicSignals (need to deserialize from strings)
-  // U: [splitToRegisters(U.x) as bigint[], splitToRegisters(U.y) as bigint[]];
 
   return false;
 }
